# Optimize resizeNearestNeighbor CPU