#### Description
<!--
Please describe the pull request here.
Also, if this is an issue/bug fix, please add the issue link for reference here.
-->
This PR optimizes the performance of the CPU implementation of `resizeNearestNeighbor`.

Test code:
```
    tf.setBackend("cpu");
    const ns = [64, 128, 256, 512];
    const RUNS = 100;

    ns.forEach(n => {
      const t = tf.randomUniform([8, n, n, 3]);
      const newSize = [n * 3 / 2, n * 3 / 2];
	// Warmup.
      tf.image.resizeNearestNeighbor(t, newSize, false);

      let res = null;
      const start = performance.now();
      for (let i = 0; i < RUNS; i++) {
        res = tf.image.resizeNearestNeighbor(t, newSize, false);
      }
      res.dataSync();
      const elapsed = (performance.now() - start) / RUNS;
      console.log(`N: ${n}: ${elapsed.toFixed(2)} ms`);
	});
```

Results before optimization:
```
N: 64: 57.66 ms
N: 128: 193.83 ms
N: 256: 717.71 ms
N: 512: 2732.96 ms
```

Results after optimization:
```
N: 64: 1.32 ms
N: 128: 5.18 ms
N: 256: 21.12 ms
N: 512: 127.87 ms
```

---
<!-- Please do not delete this section -->
##### For repository owners only:

Please remember to apply all applicable tags to your pull request.
Tags: FEATURE, BREAKING, BUG, PERF, DEV, DOC, SECURITY

For more info see: https://github.com/tensorflow/tfjs/blob/master/DEVELOPMENT.md

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/tensorflow/tfjs-core/1316)
<!-- Reviewable:end -->